plugin.hpp is intended for shared plugin-wide symbols so Common.hpp is redundant.
