+++ src/DualAtenuverter.cpp
@@ -1,5 +1,4 @@
 #include "plugin.hpp"
-#include "Common.hpp"
 
 struct DualAtenuverter : Module {
 	enum ParamIds {
