+++ src/Kickall.cpp
@@ -1,5 +1,4 @@
 #include "plugin.hpp"
-#include "Common.hpp"
 #include "ChowDSP.hpp"
